requirements.txt
html-testRunner==1.2.1
requests==2.27.1
Flask==3.0.0
Werkzeug==3.0.1
pymysql==1.0.2
flask_cors==3.0.10
bcrypt==3.2.0
mysql-connector-python==8.0.28